Router: extract rules to Splits, use class lookup
A profiler run revealed that `getSplits` is the largest bottleneck of
performance, despite being run only once, almost double that of the part
which was previously thought to be the more expensive: BestFirstSearch.

Let's refactor away from using a massive `match` to several lookups by
the token class.
