+++ scalafmt-core/shared/src/main/scala/org/scalafmt/internal/Constants.scala
@@ -1,0 +1,12 @@
+package org.scalafmt.internal
+
+object Constants {
+  val ShouldBeNewline = 100000
+  val ShouldBeSingleLine = 30
+  val BinPackAssignmentPenalty = 10
+  val SparkColonNewline = 10
+  val BracketPenalty = 20
+  val ExceedColumnPenalty = 1000
+  // Breaking a line like s"aaaaaaa${111111 + 22222}" should be last resort.
+  val BreakSingleLineInterpolatedString = 1000 * ExceedColumnPenalty
+}
